backend/requirements.txt
django
djangorestframework
django-cors-headers
psycopg2-binary
dotenv
djangorestframework-simplejwt
stripe